+++ Elements/src/Geometry/Polygon.cs
@@ -1126,7 +1126,24 @@ namespace Elements.Geometry
                 var v = p[i];
                 converted[i] = new Vector3(v.X / scale, v.Y / scale);
             }
-            return new Polygon(converted);
+            try
+            {
+                return new Polygon(converted);
+            }
+            catch (ArgumentException e)
+            {
+                // Often, the polygons coming back from clipper will have self-intersections, in the form of lines that go out and back. 
+                // here we make a last-ditch attempt to fix this and construct a new polygon. 
+                var cleanedVertices = Vector3.AttemptPostClipperCleanup(converted);
+                try
+                {
+                    return new Polygon(cleanedVertices);
+                }
+                catch (Exception e2)
+                {
+                    throw new Exception("Unable to clean up bad polygon resulting from a polygon boolean operation.");
+                }
+            }
         }
 
         public static IList<Polygon> Reversed(this IList<Polygon> polygons)