+++ Elements/test/ProfileTests.cs
@@ -221,5 +221,44 @@ namespace Elements.Tests
             Model.AddElements(difference.Select(d => (new Floor(d, 0.1))));
             Assert.Equal(3, difference.Count);
         }
+
+        [Fact]
+        public void SplitProfiles()
+        {
+            Name = "Profile_Split";
+            var profile = new Profile(
+                Polygon.Rectangle(20, 20),
+                new[] {
+                    Polygon.Rectangle(5, 5),
+                    Polygon.Rectangle(new Vector3(8, -8), new Vector3(9, 9))
+                    },
+                Guid.NewGuid(), null);
+            // this one crosses all the way thru, and should split the profile.
+            var polyline1 = new Polyline(new[] {
+                new Vector3(-21, 0),
+                new Vector3(0, 5),
+                new Vector3(21, -4),
+            });
+            // this one doesn't and should be ignored by the split
+            var polyline2 = new Polyline(new[] {
+                new Vector3(-21, -6),
+                new Vector3(0, -8),
+            });
+            //this one is totally internal and should also be ignored by the split
+            var polyline3 = new Polyline(new [] {
+                new Vector3(2, -4),
+                new Vector3(-2, -7)
+            });
+            Model.AddElement(polyline1);
+            Model.AddElement(polyline2);
+            Model.AddElement(polyline3);
+            Model.AddElements(profile.ToModelCurves());
+            var split = Elements.Geometry.Profile.Split(new[] { profile }, new[] { polyline1, polyline2, polyline3 });
+            var random = new Random(4);
+            Model.AddElements(split.SelectMany(s => s.ToModelCurves()));
+            Model.AddElements(split.Select(s => new Floor(s, 0.1, new Transform(), random.NextMaterial())));
+            Assert.Equal(2, split.Count);
+            Assert.Equal(20 * 20 - 5 * 5 - 17, split.Sum(s => s.Area()), 4);
+        }
     }
 }